[feat] Add simulate endpoint in controller

+++ app/models/dtos/AmortizationRequest.py
@@ -3,4 +3,4 @@ from pydantic import BaseModel
 class AmortizationRequest(BaseModel):
     monto: float
     tasa_anual: float
-    plazo_meses: float
+    plazo_meses: int
